Version_2.1-16-5-25-22-53

+++ src/main/java/Finance/Bank_System/entities/Account/CustomerEntityAccount.java
@@ -39,7 +39,7 @@ public class CustomerEntityAccount {
     private String accountIban;
 
     @Column(name = "account_balance")
-    private int accountBalance;
+    private double accountBalance;
 
     @Column(name = "account_token", length = 36, unique = true)
     private String accountToken;
